# src/backend/src/api/routes/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swimmer reaction time meter API",
        "version": "1.0.0",
        "title": "Swagger swimmer reaction time meter"
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/swimmer/{swimmerId}": {
            "delete": {
                "description": "Deletes swimmer related to that coach id",
                "operationId": "deleteSwimmer",
                "parameters": [
                    {
                        "name": "coachId",
                        "in": "path",
                        "description": "coach Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok"
                    }
                }
            },
            "patch": {
                "description": "Updates a swimmer data",
                "parameters": [
                    {
                        "name": "coachId",
                        "in": "path",
                        "description": "coach Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "swimmer",
                        "description": "The updated swimmer info",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/swimmerInput"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Returns the updated coach data",
                        "schema": {
                            "$ref": "#/definitions/swimmerOutput"
                        }
                    }
                }
            }
        },
...
        "/coach/{coachId}": {
            "get": {
                "description": "Returns a coach by id",
                "operationId": "getCoachById",
                "parameters": [
                    {
                        "name": "coachId",
                        "in": "path",
                        "description": "coach Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Returns the coach data",
                        "schema": {
                            "$ref": "#/definitions/coachOutput"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a coach and its swimmers",
                "operationId": "deleteCoach",
                "parameters": [
                    {
                        "name": "coachId",
                        "in": "path",
                        "description": "coach Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Ok"
                    }
                }
            },
            "patch": {
                "description": "Updates a coach data",
                "parameters": [
                    {
                        "name": "coachId",
                        "in": "path",
                        "description": "coach Id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "coach",
                        "description": "The updated coach info",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/coachInput"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Returns the updated coach data",
                        "schema": {
                            "$ref": "#/definitions/coachOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "coachOutput": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "coachInput": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "swimmerInput": {
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "swimmerOutput": {
            "properties": {
                "name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
    },
...
}